# openapi/openrag-lab.yaml
...
components:
  schemas:
    Body_ingest_document_api_documents_ingest_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      required:
      - file
      title: Body_ingest_document_api_documents_ingest_post
      type: object
...
    DeleteDocumentOut:
      properties:
        deleted_chunks:
          title: Deleted Chunks
          type: integer
        filename:
          title: Filename
          type: string
        stored_filename:
          title: Stored Filename
          type: string
      required:
      - filename
      - stored_filename
      - deleted_chunks
      title: DeleteDocumentOut
      type: object
    DocumentListOut:
      properties:
        files:
          items:
            $ref: '#/components/schemas/DocumentOut'
          title: Files
          type: array
        total:
          title: Total
          type: integer
      required:
      - total
      - files
      title: DocumentListOut
      type: object
    DocumentOut:
      properties:
        created_at:
          title: Created At
          type: string
        display_name:
          title: Display Name
          type: string
        id:
          title: Id
          type: string
        mimetype:
          title: Mimetype
          type: string
        openrag_document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Openrag Document Id
        size_bytes:
          title: Size Bytes
          type: integer
        status:
          title: Status
          type: string
        stored_filename:
          title: Stored Filename
          type: string
        updated_at:
          title: Updated At
          type: string
        uploaded_by:
          title: Uploaded By
          type: string
      required:
      - id
      - display_name
      - stored_filename
      - mimetype
      - size_bytes
      - uploaded_by
      - status
      - created_at
      - updated_at
      title: DocumentOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: OpenRAG Lab API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/documents:
    get:
      description: List the documents the caller's tenant has registered.
      operationId: list_documents_api_documents_get
      parameters:
      - in: query
        name: tenant_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Documents
      tags:
      - documents
  /api/documents/ingest:
    post:
      description: 'Store an uploaded file in OpenRAG under the tenant''s namespace.


        The document is registered only after OpenRAG reports the ingestion task

        finished, so a failed ingest leaves no registry entry behind.'
      operationId: ingest_document_api_documents_ingest_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_ingest_document_api_documents_ingest_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Ingest Document
      tags:
      - documents
  /api/documents/{filename}:
    delete:
      description: Delete a registered document by its display name.
      operationId: delete_document_api_documents__filename__delete
      parameters:
      - in: path
        name: filename
        required: true
        schema:
          title: Filename
          type: string
      - in: query
        name: tenant_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteDocumentOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete Document
      tags:
      - documents
...